requirement.txt
numpy==2.0.2
tensorflow==2.18.0
matplotlib==3.9.2
scipy==1.14.1
scikit-learn==1.5.2
umap-learn==0.5.7
POT==0.9.5
ott-jax==0.4.9
scanpy==1.10.3
jax==0.4.35
jaxlib==0.4.35
